LayoutTests/webaudio/OfflineAudioContext/offlineaudiocontext-leak.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description("Makes sure that the OfflineAudioContext objects are not leaking.");
jsTestIsAsync = true;

const instancesToCreate = 100;

function allocateContexts() {
    let contexts = [];
    for (let i = 0; i < instancesToCreate; i++)
        contexts.push(new OfflineAudioContext(2, 1, 44100));
    shouldBeTrue("internals.numberOfBaseAudioContexts() >= instancesToCreate");
}

allocateContexts();
gc();
setTimeout(() => {
    gc();
    shouldBeTrue("internals.numberOfBaseAudioContexts() < instancesToCreate");
    finishJSTest();
}, 0);
</script>
</body>
</html>
